0.12.00.03 (for Melvor 0.12.2)
Added more farm planting modes.  Thank you Daedalus, for the inspiration!
- Seeds, herbs and trees can be configured to be planted by base XP, mastery, farming level or to replant what's there.
- Each mode can be sorted ascending or descending. Though this is ignored for replanting.

Because of this, the configuration for planting fields has changed slightly.  I hope it's still intuitive enough for changing the settings to your liking.

The script will now try and termporarily equip the farming skillcape if you own it.  It needs to be in your inventory, and you need at least 1 bank spot available for this to work.  While in combat, the script will opt to buy any needed compost instead of messing with your build. (Thank you Virus, for the suggestion!)  If the farming skillcape is in an equipment profile you're not using, the script won't find it.
After use, your original (skill)cape will be put back.

Some very minor tweaks and adjustments.